+++ summoner-cli/src/Summoner.hs
@@ -1,3 +1,11 @@
+{- |
+Copyright: (c) 2017-2019 Kowainik
+SPDX-License-Identifier: MPL-2.0
+Maintainer: Kowainik <xrom.xkov@gmail.com>
+
+Main module that reexports all library components of the @summoner@.
+-}
+
 module Summoner
        ( module Summoner
        ) where
@@ -6,12 +14,14 @@ import Summoner.Ansi as Summoner
 import Summoner.CLI as Summoner
 import Summoner.Config as Summoner
 import Summoner.CustomPrelude as Summoner
+import Summoner.Decision as Summoner
 import Summoner.Default as Summoner
 import Summoner.GhcVer as Summoner
 import Summoner.License as Summoner
 import Summoner.Project as Summoner
 import Summoner.Question as Summoner
 import Summoner.Settings as Summoner
+import Summoner.Source as Summoner
 import Summoner.Template as Summoner
 import Summoner.Text as Summoner
 import Summoner.Tree as Summoner