+++ summoner-cli/src/Summoner/Ansi.hs
@@ -1,4 +1,10 @@
--- | This module contains functions for colorful printing into terminal.
+{- |
+Copyright: (c) 2017-2019 Kowainik
+SPDX-License-Identifier: MPL-2.0
+Maintainer: Kowainik <xrom.xkov@gmail.com>
+
+This module contains functions for colorful printing into terminal.
+-}
 
 module Summoner.Ansi
        ( Color (..)