+++ summoner-cli/src/Summoner/Default.hs
@@ -1,4 +1,10 @@
--- | This module contains some default values to use.
+{- |
+Copyright: (c) 2017-2019 Kowainik
+SPDX-License-Identifier: MPL-2.0
+Maintainer: Kowainik <xrom.xkov@gmail.com>
+
+This module contains some default values to use.
+-}
 
 module Summoner.Default
        ( defaultGHC