+++ summoner-cli/src/Summoner/Settings.hs
@@ -1,3 +1,11 @@
+{- |
+Copyright: (c) 2017-2019 Kowainik
+SPDX-License-Identifier: MPL-2.0
+Maintainer: Kowainik <xrom.xkov@gmail.com>
+
+Complete settings required for the project creation.
+-}
+
 module Summoner.Settings
        ( Settings (..)