+++ summoner-cli/src/Summoner/Template.hs
@@ -1,4 +1,10 @@
--- | This module contains functions for stack template creation.
+{- |
+Copyright: (c) 2017-2019 Kowainik
+SPDX-License-Identifier: MPL-2.0
+Maintainer: Kowainik <xrom.xkov@gmail.com>
+
+This module contains functions for the Haskell project template creation.
+-}
 
 module Summoner.Template
        ( createProjectTemplate