+++ summoner-cli/src/Summoner/Template/Cabal.hs
@@ -1,6 +1,12 @@
 {-# LANGUAGE QuasiQuotes #-}
 
-{- | `.cabal` file template. -}
+{- |
+Copyright: (c) 2017-2019 Kowainik
+SPDX-License-Identifier: MPL-2.0
+Maintainer: Kowainik <xrom.xkov@gmail.com>
+
+@.cabal@ file template.
+-}
 
 module Summoner.Template.Cabal
        ( cabalFile