+++ summoner-cli/src/Summoner/Template/Haskell.hs
@@ -1,5 +1,13 @@
 {-# LANGUAGE QuasiQuotes #-}
 
+{- |
+Copyright: (c) 2017-2019 Kowainik
+SPDX-License-Identifier: MPL-2.0
+Maintainer: Kowainik <xrom.xkov@gmail.com>
+
+Templates for generated Haskell source code files.
+-}
+
 module Summoner.Template.Haskell
        ( haskellFiles
        ) where