+++ summoner-cli/src/Summoner/Tree.hs
@@ -1,5 +1,13 @@
 {-# LANGUAGE ViewPatterns #-}
 
+{- |
+Copyright: (c) 2017-2019 Kowainik
+SPDX-License-Identifier: MPL-2.0
+Maintainer: Kowainik <xrom.xkov@gmail.com>
+
+Data type for representing filesystem structure via tree.
+-}
+
 module Summoner.Tree
        ( TreeFs (..)
        , traverseTree