+++ summoner-tui/src/Summoner/Tui/Form.hs
@@ -1,4 +1,10 @@
--- | Form layout and form fields data types.
+{- |
+Copyright: (c) 2018-2019 Kowainik
+SPDX-License-Identifier: MPL-2.0
+Maintainer: Kowainik <xrom.xkov@gmail.com>
+
+Form layout and form fields data types.
+-}
 
 module Summoner.Tui.Form
        ( SummonForm (..)